Changed to use Long instead of Integer. :(

+++ src/main/java/be/tomcools/tombot/model/FacebookMessageContent.java
@@ -2,6 +2,6 @@ package be.tomcools.tombot.model;
 
 public class FacebookMessageContent {
     public String mid;
-    public Integer seq;
+    public Long seq;
     public String text;
 }
